Change Transformation on Mesh to a vector of Transformations
This attribute logic and usage was changed because the Mesh will now be treated as a *primitive* instead of an object on a scene. It is important to distinguish between an object drawn in a scene and a primitive.

For example: the table is currently using 5 cube meshes, and all the cubes have the same exact vertices positions and indices data, the only thing that differentiates them are the transformation that are done on them, so they change scale, position and rotation. So, instead of repeating the same information 5 times, wasting buffer transfer time and unnecessary space, why not only store the information of a single *primitive* cube and 5 transformations to it and iterate on those transformations, drawing it 5 times, each with a different transformation?

That is the change being made in this commit. What is left to do:
1) Rewrite the createTable( ) function to take advantage of this change;
2) Maybe change in the future how color is passed to the Mesh because, the way it is now, the color is fixed on the Mesh. So, if there are multiple transformations of the primitive to be drawn, all of them will have the same color.