Add Arcball rotation interaction and GLM libs
The initial idea was to adapt the code from https://en.wikibooks.org/wiki/OpenGL_Programming/Modern_OpenGL_Tutorial_Arcball to use the current libs and objects that I'm using on the project. But, to not waste much time adapting the code, I decided to import the whole GLM into the project to use it in the Arcball calculation.

Add GLM version 0.9.9.8.

Remove idle rotation on Y axis animation and add Arcball world rotation using the left mouse button click and hold interaction in its place.