Fix normal vector and ViewDirection calculations on shader
The normal vector needs to follow the transformations made on the models! That's why now the model transformation matrix is sent to the vertex shader.

The viewDir vector was using lightPosition in its calculation but it needs to use the camera position instead! The lightning in the scene is directional, not spot, so lightPosition is not being used.